Add new user models and update EF context
This commit introduces new entity models `DName` and `UserData` to the `Test_Application` project. It updates the context classes (`User.Context.cs` and `User.Context.tt`) and modifies the EDMX files to reflect the new database schema. The `FormValidationController.cs` is enhanced with a new action method for form submissions. Additionally, the project file (`Test Application.csproj`) is updated to accommodate the new EF structure, and the connection strings in `Web.config` are modified to include new database connections for the `ApplicationDB` context. These changes improve data handling capabilities within the application.